Added entering/exiting CharacterInfoScreen
You can enter and leave character info screen. Only displays character
image right now
