+++ app/src/main/java/ultrapoulet/wifeygame/battle/BattleSelectScreen.java
@@ -3,7 +3,6 @@ package ultrapoulet.wifeygame.battle;
 import android.graphics.Color;
 import android.graphics.Paint;
 
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
@@ -12,11 +11,11 @@ import ultrapoulet.androidgame.framework.Graphics;
 import ultrapoulet.androidgame.framework.Input.TouchEvent;
 import ultrapoulet.androidgame.framework.Screen;
 import ultrapoulet.wifeygame.Assets;
+import ultrapoulet.wifeygame.PartySelectScreen;
 import ultrapoulet.wifeygame.WifeyCharacter;
 import ultrapoulet.wifeygame.battle.enemyai.BasicMagicEnemyAI;
 import ultrapoulet.wifeygame.battle.enemyai.BasicPhysicalEnemyAI;
 import ultrapoulet.wifeygame.battle.enemyai.OriginalBossAI;
-import ultrapoulet.wifeygame.gamestate.Party;
 
 /**
  * Created by John on 4/26/2016.
